# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Hotel Group Internal Procurement MVP API
  version: 0.3.0
  description: >
    Formal MVP API contract covering P0 controls and Phase 1 supplier management,
    procurement request and project initiation. This contract keeps the MVP boundary:
    no CA, e-signature, bid-file encryption/decryption, trusted timestamp,
    tamper-proof evidence, opening hall, or real external system integration.
servers:
  - url: http://127.0.0.1:3000
security:
  - MockUserHeader: []
components:
  securitySchemes:
    MockUserHeader:
      type: apiKey
      in: header
      name: x-mock-user-id
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            auditLogId:
              type: string
  responses:
    Ok:
      description: Request succeeded.
...
    PermissionDenied:
      description: Request rejected by backend policy and recorded with auditLogId when applicable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
paths:
...
  /api/archive-supplement-requests/{requestId}/apply:
    post:
      tags: [Archive]
      summary: Apply approved archive supplement metadata
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "400":
          description: Supplement request must be approved
  /api/archive-items/{itemId}/update:
    post:
      tags: [Archive]
      summary: Archive item direct update guard
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "403":
          $ref: "#/components/responses/PermissionDenied"
...